+++ triplea/config/settings.py
@@ -7,9 +7,9 @@ from pydantic_settings import BaseSettings
 from dotenv import load_dotenv
 
 import tomli
+
 # https://stackoverflow.com/questions/67085041/how-to-specify-version-in-only-one-place-when-using-pyproject-toml # noqa: E501
 import importlib.metadata
-
 
 
 # Project Directories
@@ -31,8 +31,8 @@ try:
         pyproject = tomli.load(f)
         version = pyproject["tool"]["poetry"]["version"]
 except Exception:
-    # This is in Package 
-    # version = importlib.metadata.version(__package__ or __name__)
+    # This is in Package
+    #  version = importlib.metadata.version(__package__ or __name__)
     version = importlib.metadata.version("triplea")
 
 
@@ -89,11 +89,9 @@ class Settings(BaseSettings):
         "AAA_FULL_TEXT_STRING_DIRECTORY", "Directory"
     )
 
-    AAA_LLM_TEMPLATE_FILE : Optional[str] = os.getenv(
+    AAA_LLM_TEMPLATE_FILE: Optional[str] = os.getenv(
         "AAA_LLM_TEMPLATE_FILE",
-        os.path.join(
-            ROOT, "service", "llm", "llm_profile_template_sample.json"
-        )
+        os.path.join(ROOT, "service", "llm", "llm_profile_template_sample.json"),
     )
 
     # class Config: